Chore(BE-8): Added some fixes to templateController and templateRoutes

+++ server/controllers/templateController.js
@@ -98,11 +98,18 @@ const deleteCoverLetter = async (req, res) => {
 	if (!mongoose.Types.ObjectId.isValid(templateId))
 		return res.status(404).json({ message: "This user id is not valid!" });
 
-	const template = await Template.findByIdAndDelete({ id: templateId });
+	const checkIfTemplateExists = await Template.find({}).count();
 
-	return res.status(StatusCodes.OK).json({
-		message: "Cover Letter deleted successfully",
-	});
+	if (!checkIfTemplateExists < 1) {
+		const template = await Template.findByIdAndDelete({ id: templateId });
+		return res.status(StatusCodes.OK).json({
+			message: `Cover Letter deleted with the id ${templateId} successfully`,
+		});
+	} else {
+		return res.status(StatusCodes.NOT_FOUND).json({
+			message: `Cover Letter with the id ${templateId} does not exist`,
+		});
+	}
 };
 
 module.exports = {
